requirements.txt
yfinance==0.2.36
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1 